fix(imitate): fix imitate, should not add listener immediately
Fix imitate method to mitigate issue #5. Now we must use xs.createMimic() to get a MimicStream, and
the MimicStream is the only type that has the imitate() method. Adding a listener to the MimicStream
literally just adds that listener to the imitated stream.

Fixes issue #5.

BREAKING CHANGE:
imitate() method on Stream removed. New type introduced: MimicStream,
which can be created through xs.createMimic(). A MimicStream has the
method imitate(), which has the same API as before, but imitate does not
trigger any Stream/Producer to start.