feat(core): improve signature of operators regarding types (#43)
This commit changes core operators to return either: (1) Stream only, (2) Stream or MemoryStream,
(3) MemoryStream only. Operators like fold() and startWith() now always return a MemoryStream, while
operators like map(), take(), and endWhen() return the same type that the source/instance stream
was. E.g. MemoryStream mapped will return a MemoryStream.

BREAKING CHANGE:
debug() now returns a MemoryStream if the input was also a MemoryStream.
endWhen() now returns a MemoryStream if the input was also a MemoryStream.
fold() now returns always a MemoryStream, not Stream.
imitate() only works on conventional Stream, will throw error on
MemoryStream.
map() now returns a MemoryStream if the input was also a MemoryStream.
mapTo() now returns a MemoryStream if the input was also a MemoryStream.
replaceError() now returns a MemoryStream if the input was also a MemoryStream.
startWith() now returns always a MemoryStream, not Stream.
take() now returns a MemoryStream if the input was also a MemoryStream.