fix(imitate): fix cyclic execution leak, and refactor
Rewrite MimicStream with a different approach. Adds a hidden listener field (_hil) to Stream, which
is a special listener which does not trigger start nor stop when added/removed. imitate just adds
the MimicStream as a hidden listener on the target Stream, and MimicStream has its own listeners, it
does not redirect them to the target. This new approach breaks the usage of imitate for
Directed-Acyclic-Graphs, but we anyway intend MimicStream to be used _only_ to implement a cycle in
the stream graph. So this commit removes some tests.

Fixes bug #51, deprecates/invalidates issue #49.