fix(imitate): fix cyclic execution leaks
Rewrite imitate() with a new approach that redirects every listener of a mimic stream directly to
the target stream, and then does graph traversal to check if an isolated cycle exists in the stream
graph, whenever we remove a listener and the stream is about to have just one listener left.

Fixes #51.