fix(merge): support union types
Allows xs.merge(streamOfNumbers, streamOfStrings) to return a Stream of numbers or strings, as the return type in TypeScript.

Modifis static `merge` method to implement a new interface:
MergeSignature. This interface is constructed similarly to the
CombineSignature interface. Export the new interface.

This change allows consumers invoking the static method to declare types
and leverage TypeScript to verify type safety when merging streams. The
change brings `merge` in line with other static methods which already
include this feature.

PR #82.

Should resolve staltz/xstream#80.