fix(flatten): when same inner stream, restart
Fix flatten operator so that if the new inner stream is the same as the previous inner stream, it
synchronously stops the inner stream before subscribing to it, which will cause a synchronous
(re)start. This is important to maintain referential transparency when the inner may be defined
either from the closure or internally in the map project function. Both cases should be the same.

Closes issue #90.