fix(flattenSequentially): stop execution of interrupted inner streams
Previously, inner streams flattened through flattenSequentially would continue
to execute even when the outer result stream was stopped. This commit fixes
this so that no leaky executions are allowed. If the flat result stream stops,
so does the inner stream.

BREAKING CHANGE:
![maybe won't](https://img.shields.io/badge/will%20it%20affect%20me%3F-maybe%20won't-yellowgreen.svg)
This is likely a breaking change for people using flattenSequentially,
specially given that xstream core was fixed so that errors are not swallowed.
Most flattenSequentially code should still work, but upgrade carefully anyway.