fix(extra): Simplify extra operator generics
Move the generic parameters of the extra operators to support better generic parameter
inferencing in Typescript 2.4 while continuing to pass tests in Typescript 2.1. Delay is
the only extra operator of this form skipped in this commit because it is the only one
that didn't pass Typescript inferencing compile (in imitate tests) in Typescript 2.1
(though it builds in Typescript 2.4).

Resolves #202