fix(imitate): use null for unset this._target
The unset this._target used to be the constant object NO but it seems
more reliable to use null here. In some rare cases (maybe with garbage
collection?) the NO object in this._target may differ from the global
NO object. The NO object is useful for streams that emit null or
undefined, but not useful for this._target so it seems better to just
use null here.