fix(throttle): clear internal interval when unsubscribed
When a throttled stream is unsubscribed, its internal throttle interval
is not cleared, and the reference to its ID is removed, meaning it's
also not cleared at the end of the timeout.

This is essentially a memory leak, where the interval will live on in
the event loop forever without being cleared.

This commit adds a call to ThrottleOperator#clearInterval() in its _stop
method, clearing the interval and freeing it from the event loop.